+++ GeomLine.cpp
@@ -4,7 +4,7 @@ void GeomLine::accept(Visitor* v) {
 	v->visit(this);
 }
 
-GeomLine::GeomLine(int x1, int y1, int x2,int y2) {
+GeomLine::GeomLine(float x1, float y1, float x2, float y2) {
 	p1 = new Point(x1, y1);
 	p2 = new Point(x2, y2);
 	points.push_back(p1);
