+++ GeomPolygone.cpp
@@ -4,10 +4,10 @@ void GeomPolygone::accept(Visitor* v) {
 	v->visit(this);
 }
 
-GeomPolygone::GeomPolygone(vector<int>& pts) {
+GeomPolygone::GeomPolygone(vector<float>& pts) {
 	for (int i = 0; i < pts.size(); i += 2) {
-		int x = pts[i];
-		int y = pts[i + 1];
+		float x = pts[i];
+		float y = pts[i + 1];
 		points.push_back(new Point(x, y));
 	}
 };
